# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AssessmentInputEnvironment = typeof AssessmentInputEnvironment[keyof typeof AssessmentInputEnvironment];


export const AssessmentInputEnvironment = {
  demo: 'demo',
  local: 'local',
  authorized: 'authorized',
} as const;

export interface AssessmentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  target: string;
  environment: AssessmentInputEnvironment;
  /** @minLength 1 */
  authorization: string;
}

export type AssessmentStatus = typeof AssessmentStatus[keyof typeof AssessmentStatus];


export const AssessmentStatus = {
  in_progress: 'in_progress',
  complete: 'complete',
  draft: 'draft',
} as const;

export type AssessmentMode = typeof AssessmentMode[keyof typeof AssessmentMode];


export const AssessmentMode = {
  simulation: 'simulation',
  local_fixture: 'local_fixture',
  authorized_target: 'authorized_target',
} as const;

export type Assessment = AssessmentInput & {
  id: string;
  status: AssessmentStatus;
  mode: AssessmentMode;
  createdAt: string;
  progress: number;
};

export type CoverageStatus = typeof CoverageStatus[keyof typeof CoverageStatus];


export const CoverageStatus = {
  not_assessed: 'not_assessed',
  in_progress: 'in_progress',
  candidate: 'candidate',
  confirmed: 'confirmed',
  verified: 'verified',
} as const;

export interface Coverage {
  area: string;
  status: CoverageStatus;
  checks: number;
}

export type SurfaceComponentSource = typeof SurfaceComponentSource[keyof typeof SurfaceComponentSource];


export const SurfaceComponentSource = {
  reference_model: 'reference_model',
  discovered: 'discovered',
} as const;

export interface SurfaceComponent {
  id: string;
  name: string;
  purpose: string;
  kind: string;
  source: SurfaceComponentSource;
  interfaces?: string[];
}

export type AssessmentDetail = Assessment & {
  coverage: Coverage[];
  components: SurfaceComponent[];
};

export interface Activity {
  id: string;
  type: string;
  title: string;
  detail: string;
  origin: string;
  time: string;
}

export interface Dashboard {
  openAssessments: number;
  candidates: number;
  confirmed: number;
  verified: number;
  progress: number;
  activity: Activity[];
  nextAction: string;
}

export interface ValidationCase {
  id: string;
  identity: string;
  owner: string;
  policy: string;
  expected: string;
  observed: string;
  vulnerable: boolean;
}

export interface ValidationInput {
  caseId: string;
}

export type ValidationRunStatus = typeof ValidationRunStatus[keyof typeof ValidationRunStatus];


export const ValidationRunStatus = {
  passed: 'passed',
  failed: 'failed',
  inconclusive: 'inconclusive',
} as const;

export interface ValidationRun {
  id: string;
  caseId: string;
  status: ValidationRunStatus;
  assertions: string[];
  request: string;
  response: string;
  origin: string;
  createdAt: string;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type FindingState = typeof FindingState[keyof typeof FindingState];


export const FindingState = {
  candidate: 'candidate',
  confirmed: 'confirmed',
  rejected: 'rejected',
  inconclusive: 'inconclusive',
} as const;

export interface Finding {
  id: string;
  title: string;
  category: string;
  component: string;
  severity: FindingSeverity;
  state: FindingState;
  origin: string;
  updatedAt: string;
}

export interface Evidence {
  id: string;
  findingId: string;
  testId: string;
  runId: string;
  environment: string;
  origin: string;
  timestamp: string;
  revision: string;
  redaction: string;
  request: string;
  response: string;
  assertions: string[];
}

export type VerificationStatus = typeof VerificationStatus[keyof typeof VerificationStatus];


export const VerificationStatus = {
  verified_fixed: 'verified_fixed',
  still_reproducible: 'still_reproducible',
  inconclusive: 'inconclusive',
} as const;

export interface Verification {
  id: string;
  findingId: string;
  status: VerificationStatus;
  assertions: string[];
  createdAt: string;
  origin: string;
}

export type FindingDetail = Finding & {
  description: string;
  rationale: string;
  prerequisites: string;
  expected: string;
  actual: string;
  steps: string[];
  cause: string;
  impact: string;
  remediation: string;
  evidence: Evidence[];
  verification: Verification[];
};

export type RemediationStatus = typeof RemediationStatus[keyof typeof RemediationStatus];


export const RemediationStatus = {
  proposed: 'proposed',
  applied: 'applied',
  ready_for_retest: 'ready_for_retest',
} as const;

export interface Remediation {
  findingId: string;
  status: RemediationStatus;
  message: string;
}

export type ReportInputFormat = typeof ReportInputFormat[keyof typeof ReportInputFormat];


export const ReportInputFormat = {
  html: 'html',
  json: 'json',
  sarif: 'sarif',
  pdf: 'pdf',
} as const;

export interface ReportInput {
  format: ReportInputFormat;
}

export interface Report {
  id: string;
  format: string;
  status: string;
  disclosure: string;
  generatedAt: string;
}

export type IntegrationStatus = typeof IntegrationStatus[keyof typeof IntegrationStatus];


export const IntegrationStatus = {
  connected: 'connected',
  not_connected: 'not_connected',
  unavailable: 'unavailable',
  error: 'error',
} as const;

export interface Integration {
  name: string;
  purpose: string;
  status: IntegrationStatus;
  mode: string;
}

export interface Workspace {
  product: string;
  assessment: Assessment;
  mode: string;
  integrations: Integration[];
}