# lib/api-zod/src/generated/types/assessmentDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { Assessment } from './assessment';
import type { Coverage } from './coverage';
import type { SurfaceComponent } from './surfaceComponent';

export type AssessmentDetail = Assessment & {
  coverage: Coverage[];
  components: SurfaceComponent[];
};