# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputEnvironment } from './assessmentInputEnvironment';

export interface AssessmentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  target: string;
  environment: AssessmentInputEnvironment;
  /** @minLength 1 */
  authorization: string;
}