# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';

export interface Dashboard {
  openAssessments: number;
  candidates: number;
  confirmed: number;
  verified: number;
  progress: number;
  activity: Activity[];
  nextAction: string;
}