# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';
import type { FindingState } from './findingState';

export interface Finding {
  id: string;
  title: string;
  category: string;
  component: string;
  severity: FindingSeverity;
  state: FindingState;
  origin: string;
  updatedAt: string;
}