# lib/api-zod/src/generated/types/findingDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { Evidence } from './evidence';
import type { Finding } from './finding';
import type { Verification } from './verification';

export type FindingDetail = Finding & {
  description: string;
  rationale: string;
  prerequisites: string;
  expected: string;
  actual: string;
  steps: string[];
  cause: string;
  impact: string;
  remediation: string;
  evidence: Evidence[];
  verification: Verification[];
};