# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}