# lib/api-zod/src/generated/types/integration.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { IntegrationStatus } from './integrationStatus';

export interface Integration {
  name: string;
  purpose: string;
  status: IntegrationStatus;
  mode: string;
}