# lib/api-zod/src/generated/types/remediation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { RemediationStatus } from './remediationStatus';

export interface Remediation {
  findingId: string;
  status: RemediationStatus;
  message: string;
}