# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CertaProof authorized security assessment workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputFormat } from './reportInputFormat';

export interface ReportInput {
  format: ReportInputFormat;
}